Initial `Target` C API (#14248)
* Initial: Enable a C API to build and modify a Target

* Add: `add_instruction` function

* Add: More methods to `Target` C API
- Add provisional `PropertyMap` to represent the mapping of `qargs : InstructionProperties` before adding it to the `Target`.
- `InstructionProperties`:
    - free.
- `Target`:
    - `update_instruction_prop`
    - `operation_names`
    - `physical_qubits`
    - `non_global_operation_names`.
    - `operation_names_for_qargs`.
    - `qargs_for_operation_names`.
    - `qargs`.
    - `instruction_supported`.
    - `contains_key`.

* Add: C unit test
- Add getter methods for `InstructionProperties`.
- Add `length` method for `PropertyMap`.
- Add name maping for `PropertyMap`.

* Fix: Add `IndexMap` and `Ahash` dependencies to qiskit_cext.
- Add `get` method to `Target`.
- Add missing methods to `PropertyMap`.
- Fix methods returning arrays to C by adding a a forget statement to prevent pointer from being destroyed.
- Add getter methods to the `C` api of target.
- Add helper function to parse qargs.

* Test: Add tests to update properties and get non global names.

* Add `operation_names_for_qargs` test case.
- Move `print_qargs` helper to top of module.

* Add: `qargs_for_operation_names` test case
- Add helper method to compare arrays.
- Fix print method for qargs.

* Docs: Add first wave of docstrings.

* Fix: Incorrect typing in `test_target.c`

* Fix: Add missing docstrings.
- Use correct type of pointers for outputs,

* Lint: Fix c-format

* Fix: Remove option from `qarg_gate_map`.
- Use IndexSet to preserve insertion order to avoid sorting in the C API.

* Test: Add `qargs` test.

* FIx: Add missing docstrings for remainign tests

* Fix: Type errors in MacOS

* Fix: Undo extra changes in the target.
- Sort the string arrays obtained from the target in C during tests.
- Remove `Copy` from `InstructionProperties`.
- Add helper method to sort string array outputs, coming from the Target.

* Add: more complete constructor for `Target`.
- Add getters for `description`, `dt` `min_length` `granularity`, `acquire_alignment`, and `pulse_alignment`.
- Add Target construction test.

* Fix: Clean up pointers in `qk_target_new`
- Add setters for `QkTarget` attributes.

* Fix: Add release note

* Fix: Typos in docstring

* Lint: Fix lint for C

* Fix: Use structs for collections, specify length.
- Introduce `QkTargetQargs`, `QkTargetNameList`, and `QkTargetQargsList` to expose collections of operation names and qargs to C exposing the length of the collection.

* Fix: Use ``Target`` for docstrings.

* Fix: Use the `qiskit_transpiler` crate.

* Fix: Add "qiskit-transpiler" to cbindgen.toml
- Remove incorrect usage of pointers for `qk_target_dt`.

* Fix: Incorrect import of `PhysicalQubit` in target.rs

* Fix: Remove usage of `string` to query instructions, use `StandardGate` instead.
- Introduce `Target` specific exitcodes starting at 300.
- Change methods to use `StandardGate` instead of String.

* Fix: Remove `description` from the C API.

* FIx: Remove `InstructionProperties` from the C API
- Introduce `QkInstProperties` as a C representable variant of `InstructionProperties`.
- Move `InstructionProperties` instantiation to function calls. Now you can directly send duration or error as function arguments.
- Make `PropsMap` be either initialized or `NULL`.

* Fix: Incorrect documentation usage for `qk_target_new()`

* Fix: Use NAN in one of the examples.

* Docs: Add documentation for the `Target` C API

* Lint: Qiskit header format in docs

* Fix: Use `NaN` when `dt` is not set.

* Refactor: Remove all querrying methods from Target, PropsMap.
- Remove all helper structs.
- All removed changes will be spun off to a different PR.

* Fix: Implement `From<TargetError>` for `ErrorCode`.

* Add: Separate method to add gates with fixed params.
- Since parameters are not fully supported in C as of yet, we have to separate the methods to make a clear division between how to add a gate with/without parameters, and making additions with fixed parameters fully explicit in the Target.

* Docs: Address new changes in documentation

* Fix: Add a way of copying a Target from C.

* Fix: Address review comments


* Lint: Fix formatting

* Remove: `Target::get` method.
- Will be re-added in future updates.

* Docs: Remove volatility line from `PropsMap` docs.

* Refactor: `QkPropsMap` is now `QkTargetEntry`.
- Add rust struct `TargetEntry` to better represent adding a gate to the Target from C,
- Add `SmallVec` as a dependency.
- Remove `qk_target_add_instruction_fixed` in favor of performing the checks during the creation of the `QkTargetEntry`.
- Update documentation with newer additions.

* Docs: update documentation to clearly

* Apply suggestions from code review


* Chore: Rename len methods.
- Rename `qk_target_len` to `qk_target_num_instructions`.
- Rename `qk_target_entry_len` to `qk_target_num_properties`.

* Chore: Fix documentation further
- Use `u32` for `Target` length and `min_length`.


* Docs: Add example to release note

* Chore: Remove unused error codes

* Fix: Remove conversion in vf2 vf2_layout
- Fix indentation in helper function docstrings in target.rs

* Apply suggestions from code review


* Fix: Make `qk_target_entry_add_property` fallible.
- `qk_target_entry_add_property` will throw an error code if the wrong qargs number is ever provided.
- Add comparison of error codes in unittests.

* Update docs/cdoc/qk-target-entry.rst


* Fix: Check pointer before consuming in `qk_target_add_instruction`.

* Chore: Clean up the  target tests.

* Fix: More mistakes in tests.

---------

+++ crates/transpiler/src/passes/disjoint_layout.rs
@@ -291,7 +291,7 @@ fn map_components(
 }
 
 fn build_coupling_map(target: &Target) -> Option<UnGraph<PhysicalQubit, ()>> {
-    let num_qubits = target.num_qubits.unwrap_or_default();
+    let num_qubits = target.num_qubits.unwrap_or_default() as usize;
     if target.num_qargs() == 0 {
         return None;
     }
